backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
Pillow==9.0.0
django-filter==24.3
